RAG_with_PDF_production/requirements.txt
langchain
langchain-community
llama-parse
fastembed
chromadb
langchain-groq
gradio
fastembed
faiss-cpu
langchain-cohere
unstructured
rank_bm25